Adding functions to write configuration to the config-EEPROM
This commit adds the complete EEPROM-layout for the USB2642 to the
Usb2642Eeprom-Module. Additionally it adds the functionallity to write the
EEPROM using an undocumented vendor specific command as it is done by
Microchip-Tools available from their website.

Functions shared between Usb2642Eeprom and Usb2642I2C have been moved to
ctypehelper.py.

Signed-off-by: Chris Fiege <c.fiege@pengutronix.de>