Fixing the use of the ioctl on x86 machines
The previous use of the ctypes-interface was not compatible with 32-bit
machines: The generated pointer to the data-structure and the used assert
did not match on these machines.

The new approach is using the ctypes-interface more correctly (whereby I am
still not sure if this is the best way). Additionally the length-asserts for
this struct have been made architecture-aware.

Signed-off-by: Chris Fiege <c.fiege@pengutronix.de>