usbsdmux: Adding smart guess whether to use client or direct mode
usbsdmux can run in either client or direct mode. Client mode was introduced
to use the systemd-service and is usually used as non-root user. On the
other hand direct mode only makes sense as root-user.

The usbsdmux now detects if it runs as root (UID 0) and runs in direct mode in
this case.
If it does not run as root it will default to client mode.

This should make the use of the tool more intuitive without changing the
actual user interface.

Also a new command line switch to force direct mode is introduced.

Signed-off-by: Chris Fiege <c.fiege@pengutronix.de>