usb2642i2c: move size assertions to the struct definitions
This way the assertions are only checked once as module import time instead
of once per usage. Also the assertion is near the definition of the structs
then which is where it belongs to.

Also explain that on 64 bit archs there are holes in sg_io_hdr_t and so its
size is bigger than the sum of its members. Additionally make the assert
for 64 bit vs. 32 bit arch dependent on sizeof(void *) instead of listing
a few 32 bit archs. This in turn allows to drop an import.