contrib: Add usbsdmux group
Limit the privileges of the service and give normal users write access
to the block device (for e.g. writing an image) via the following
measures:
- Create a usbsdmux user and group via sysusers.d file.
- Assign block and sg devices to the usbsdmux group via udev.
- Run the service as user and group usbsdmux.

Only if write access to the block device is required should the user
become a member of the usbsdmux group; it is not necessary for socket
communication with the service.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>