Validate sg device path and descriptor
The usbsdmux command and usbsdmux-service do not validate
the supplied device path. This is not ideal, especially for
the service, as it allows normal users to send a sequence
of weird ioctl to any device they want.

This commit adds checks for the device path and USB descriptor
values while still allowing the user to override these checks.
This fixes #26.