Remove the service/client split and access /dev/sg* directly
This change is based on the work done by sephalon, who discovered
that CAP_SYS_RAWIO is not required to perform the required ioctls
on /dev/sg* devices when the device is opened for writing.

The original suggestion was to use this fact to run the service
as a user with write access to the /dev/sg* devices belonging
to USB-SD-Muxes and grant access to the service socket to users
based on group membership.

We instead decided to grant direct access to the /dev/sg*
devices to the end-users based on group membership and
deprecate the service altogether.

Signed-Off-By: Leonard Göhrs <lgo@pengutronix.de>