commandline: 'get' must respect hardware signal priority
The previous implementation of this function did not correctly report
the off-state. Even if the device itself was off this function reported
the USB-SD-Mux to be in "dut-mode".

With this change the actual preference in hardware is respected.
In hardware the PWR_disable and DAT_disable signal has priority over the
switch-state selection.
Thus we need to check the disable-bit first. Only if disable is not set
we can be in dut- or host-mode.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>