USB-SD-Mux driver: Add support for USB-SD-Mux FAST
This change adds support for the USB-SD-Mux FAST on the device-level.
This introduces the following changes:

The `autoselect_driver()`-functions tries to probe the type of
USB-SD-Mux used and returns an appropriate driver instance.

`UsbSdMux` is now an abstract class.
The old driver now resides in `UsbSdMuxLegacy`.

The new driver lives in the `UsbSdMuxFast` class.

To add the new features to the CLI the `ArgumentParser` has been
rewritten to use subparsers instead of a positional argument.
This way we can clearly parse the additional arguments needed for the
GPIOs.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>