Drop debian packaging
Having an outdated debian packaging infrastructure in the upstream repo
is more annoying than it's worth. Given that there is also an usbsdmux
package in Debian proper, let's not interfere with that.

Additionally that makes it easier to handle our internal Debian
packaging to coexist with the official Debian package.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>