i2c_gpio: Refactor `read_register()` to be private
In c36df6096d367cbf81f38940691ef5df869d35d5 we have introduced a method
to read a register from the I2C port expander.
This method was used to read the current state of the outputs out of an
connected USB-SD-Mux.

But having this method public breaks the abstraction of the I2C-GPIO
driver. No one outside of the driver should need to care about register
addresses.

So a better way is to implement the actually used high-level functions
and expose those.

Also the method implemented the option to read more than one register in
sequence. But this wasn't use by a single caller. So we get rid of this
option.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>