Add JSON status output for all commands and errors
This commit adds JSON output to all CLI commands, if the --json switch
is set.
If the command does not output anything: Return an empty message.

It also makes sure that error messages are output in JSON, if it is an
Exception, that the CLI has already handled before.
Thus the key ``error-message`` will be present if the return code is
``!= 0``.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>