README: use python3 -m venv / python3 -m pip for installation
When the README was written the `venv` tool was still quite new and not
yet well established, so we choose to use the more conservative
virtualenv tool.

Time has passed and now `venv` is considered the best practice,
so we should advice using it in our README.

While at it also use `python3 -m pip` to call `pip`,
because it is a good way to make 100% sure that the correct `pip`
is called.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>