README: move sections around and add installation from source instructions
This is intended to get new users up to speed quicker.
For them knowing how the USB-SD-Mux works internally is certainly
interesting (at least to me it is),
but having a quick start guide on how to install the software
and use it is even more important.

This also re-structures the quickstart section so that the installation
from different sources is part of it in the form of sub-sections.
This integrates the "Distribution Packages" section better.

I was also a bit confused that none of the installation methods cloned
the git repository, so I've added one.

The "Troubleshooting" section was moved to the bottom of the page,
because that is where I would expect it to be.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>