CI: Split publish into test and production
In workflow
https://github.com/linux-automation/usbsdmux/actions/runs/11659598753/job/32460446373
we have seen that a dirty environment can be problematic for a workflow.
Here a leftover file from the publish to test.pypi.org caused the preflight
checks for the proper release to fail.

With this change we will use different jobs for both steps - thus starting
with a clean environment every time.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>