usbsdmux: sd_regs: Fix ruff error regarding the use of percent formatting
The primary goal of this commit is fixing the following `ruff`
linting error:

    usbsdmux/sd_regs.py:39:50: UP031 Use format specifiers instead of percent format
       |
    37 |             text.append(f"  {f['field']}")
    38 |         raw = f["raw"]
    39 |         text.append(f"    raw: 0b{format(raw[0], '0%db' % raw[1])} == 0x{raw[0]:0x} == {int(raw[0])}")
       |                                                  ^^^^^^ UP031
    40 |         if "enum" in f:
    41 |             text.append(f"    enum: {f['enum']} {f.get('unit', '')}".rstrip())
       |
       = help: Replace with format specifiers

While at it I had a hard time understanding what the offending code does,
so I have decided to restructure it a bit while at it.

The code path is covered by the pytest testsuite, so we can be somewhat
certain that the behaviour did not change.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>