MQTT: Print Error, but Continue if Sending Fails
If anything goes wrong during sending of statistics to MQTT we do not
want to quit the tool with a stacktrace, since this would break CI and
hinder developers from working if some part of the MQTT infrastructure
does not work.

Let's catch *all* errors during MQTT sending instead and print a simple
error message. But continue to do our job anyway.

This way an interactive user may observe the problem and can inform it -
but it does not prevent normal operation.

Fixes #93

Signed-off-by: Chris Fiege <cfi@pengutronix.de>