Treewide: Add SPDX License and Copyright Information
This change adds SPDX License and Copyright information to all files in
the tree.

Copyright information has been recovered from the git history.
As license the LGPL-2.1-or-later, as defined by COPYING, has been
assumed for most files.

For the files mentioned in REUSE.toml, that have exclusively been edited by
Pengutronix, the license was changed from LGPL-2.1-or-later to CC0-1.0.
For these files we assume that the threshold of copyrightability has not
been reached. So we do not want to derive LGPL obligations from these
snippets.

With this change the project complies with the version 3.3 of the REUSE
specification:
https://reuse.software/spec-3.3

Signed-off-by: Chris Fiege <cfi@pengutronix.de>